+++ Changes.md
@@ -11,6 +11,9 @@
 - Fix bytecode serialization of structs with missing fields, e.g., missing keys
   with maps serialized as structs
 - Remove for `arrow==35` support
+- Add `arrow=40`, `arrow=41`, `arrow=42`, `arrow=43` support
+- Add support for coercing different numeric types (use
+  `TracingOptions::default().coerce_numbers(true)`) 
 
 ## 0.7.1